requirements.txt
flask
pillow
pandas
torch
transformers